+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Skin Composer Version 53 ###
-* Updated to Stripe 1.4.1 and TextraTypist 0.6.3 
+* Updated to Stripe 1.4.1 and TextraTypist 0.7.4
+* Added new effects from TextraTypist
 * Fixed file dialog crashing on M1 Macs. Thanks Lyze!
 * Fixed floating point number parsing. Thanks H0k493!
 * Added arm64 support to work on macOS.